committing work so I be brave in refactoring this to work
