+++ src/test/java/gov/cida/cdat/db/Pojo.java
@@ -1,0 +1,7 @@
+package gov.cida.cdat.db;
+
+public class Pojo {
+	public String name;
+	public String address;
+	public String phone;
+}
